Refactor to break `Consolidate` up
