+++ source/ConsolidateCalamariPackagesTask/IPackageReference.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+
+namespace Octopus.Build.ConsolidateCalamariPackagesTask
+{
+    interface IPackageReference
+    {
+        string Name { get; }
+        string Version { get; }
+        IReadOnlyList<SourceFile> GetSourceFiles();
+    }
+}
